Complete test for camera and gallery intent
