+++ app/src/main/java/com/example/pocketinventory/ItemAddActivity.java
@@ -100,15 +100,6 @@ public class ItemAddActivity extends AppCompatActivity {
         uploadButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-               if (ContextCompat.checkSelfPermission(getApplicationContext(), android.Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-                    requestPermissions(new String[]{android.Manifest.permission.CAMERA}, 100);
-                }
-
-                // start camera activity and grab the image taken
-                Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                fileUri = getOutputMediaFileUri();
-                intent.putExtra(MediaStore.EXTRA_OUTPUT, fileUri);
-                startActivityForResult(intent, 100);
 
                 CharSequence[] items = {"Camera", "Gallery"};
                 AlertDialog.Builder builder = new AlertDialog.Builder(ItemAddActivity.this);
@@ -116,15 +107,15 @@ public class ItemAddActivity extends AppCompatActivity {
                 builder.setItems(items, (dialog, which) -> {
                     if (which == 0) {
                         // check camera permissions
-                        if (ContextCompat.checkSelfPermission(getApplicationContext(), android.Manifest.permission.CAMERA) == PackageManager.PERMISSION_DENIED) {
+                        if (ContextCompat.checkSelfPermission(getApplicationContext(), android.Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
                             requestPermissions(new String[]{android.Manifest.permission.CAMERA}, 0);
-                        } else {
+                        }
                             // start camera activity and grab the image taken
                             Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
                             fileUri = getOutputMediaFileUri();
                             intent.putExtra(MediaStore.EXTRA_OUTPUT, fileUri);
                             startActivityForResult(intent, 100);
-                        }
+
                     } else {
                         // check gallery permissions
                         if (ContextCompat.checkSelfPermission(ItemAddActivity.this, Manifest.permission.READ_MEDIA_IMAGES) == PackageManager.PERMISSION_DENIED) {
